+++ project/accounts/forms.py
@@ -15,10 +15,11 @@ from django.utils.encoding import force_bytes
 from django.utils.http import urlsafe_base64_encode
 from django.template.loader import render_to_string
 from django.utils.translation import ugettext_lazy as _
+from django.conf import settings
 
 from accounts.utils import send_email
 from .reserved_usernames import RESERVED_USERNAMES
-
+from api.models import Account
 
 User = get_user_model()
 
@@ -197,3 +198,19 @@ class RecoverUserForm(AuthRecoverUserForm):
                 recipient_list=[email],
                 html_message=html_message,
             )
+
+class UpdateAccount(forms.ModelForm):
+    """
+    Form for updating Account data
+    """
+
+    class Meta:
+        model = Account
+        fields = ["first_name", "last_name", "about_me", "profile_image", 'username', 'email']
+
+    first_name = forms.CharField(label="First Name", max_length=63, required=False)
+    last_name = forms.CharField(label="Last Name", max_length=63, required=False)
+    about_me = forms.CharField(label="About Me", max_length=511, required=False)
+    email = forms.EmailField(label="Email", disabled=True)
+    username = forms.CharField(label='Username', disabled=True)
+    profile_image = forms.ImageField(required=False)
